Canceled the commit "Remove selecting logic from InputInterpreter" of yesterday
Mario, to be sure the unit is in the selection, we need to ask the
spatial selector what is the unit coordinate in screen space. 

If we do the inverse - meaning translating the selection rectangle
cordinates to landscape space - the rectangle will be deformed due to
the camera inclination, and the unit selection will be falsed.

In a better implementation, we will translate each 4 rectangle's corners
(or even more points) from screen to landscape and compute the correct
deformed bounding zone (including some relief deformation), but it
require some more code.